Renames seeds to make room for new client table
