+++ models/seeds/02-users.js
@@ -1,0 +1,8 @@
+exports.seed = function(knex) {
+  // delete all existing entries
+  return knex('users').insert([
+    {
+      facebook_id: 10152368852405295
+    }
+  ]);
+};
